QuestionComment RestController로 변환

+++ back/noteu/src/main/java/com/noteu/noteu/question/controller/QuestionCommentController.java
@@ -5,34 +5,48 @@ import com.noteu.noteu.question.dto.request.RequestQuestionCommentDTO;
 import com.noteu.noteu.question.service.QuestionCommentService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
 
-@Controller
 @Slf4j
+@RestController
 @RequiredArgsConstructor
 @RequestMapping("/subjects/{subject-id}/questions/{question-id}")
 public class QuestionCommentController {
 
     private final QuestionCommentService questionCommentService;
 
-    @PostMapping("/question-comment")
-    public String addQuestionComment(@AuthenticationPrincipal MemberInfo memberInfo, @PathVariable("subject-id") Long subjectId,
-                                     @PathVariable("question-id") Long questionPostId, RequestQuestionCommentDTO requestQuestionCommentDTO) {
-        questionCommentService.save(requestQuestionCommentDTO, questionPostId, memberInfo.getId());
+    @PostMapping
+    public ResponseEntity<Void> addQuestionComment(@AuthenticationPrincipal MemberInfo memberInfo, @PathVariable("question-id") Long questionPostId,
+                                                   @RequestBody RequestQuestionCommentDTO requestQuestionCommentDTO) {
+        if(memberInfo == null) {
+            return ResponseEntity
+                    .status(HttpStatus.UNAUTHORIZED)
+                    .build();
+        } else {
+            questionCommentService.save(requestQuestionCommentDTO, questionPostId, memberInfo.getId());
 
-        return "redirect:/subjects/{subject-id}/questions/{question-id}";
+            return ResponseEntity
+                    .status(HttpStatus.CREATED)
+                    .build();
+        }
     }
 
-    @GetMapping("/delete/{question-commentId}")
-    public String deleteQuestionCommentById(@AuthenticationPrincipal MemberInfo memberInfo, @PathVariable("subject-id") Long subjectId,
-                                            @PathVariable("question-id") Long questionPostId, @PathVariable("question-commentId") Long questionCommentId) {
-        questionCommentService.deleteById(questionCommentId);
+    @DeleteMapping("/{question-commentId}")
+    public ResponseEntity<Void> deleteQuestionCommentById(@AuthenticationPrincipal MemberInfo memberInfo, @PathVariable("question-commentId") Long questionCommentId) {
+        if(memberInfo.getId() != questionCommentService.getById(questionCommentId).getMemberId()){
+            return ResponseEntity
+                    .status(HttpStatus.FORBIDDEN)
+                    .build();
+        } else {
+            questionCommentService.deleteById(questionCommentId);
 
-        return "redirect:/subjects/{subject-id}/questions/{question-id}";
+            return ResponseEntity
+                    .status(HttpStatus.NO_CONTENT)
+                    .build();
+        }
     }
 }
+
